restructure tests like the code base to increase alignment
