+++ multisensor_pipeline/tests/audio/test_wave.py
@@ -1,32 +1,9 @@
 from time import sleep
-import logging
 import pathlib
 
 from multisensor_pipeline.modules.audio.microphone import Microphone
 from multisensor_pipeline.modules.audio.wave import WaveFile
-from multisensor_pipeline.modules import ListSink
 from multisensor_pipeline.pipeline.graph import GraphPipeline
-
-
-def _test_microphone_input():
-    sink = ListSink()
-    try:
-        mic = Microphone(channels=1)
-        mic.add_observer(sink)
-        sink.start()
-        mic.start()
-        logging.info("recording started")
-        sleep(.2)
-    except Exception as e:
-        logging.error(e)
-    finally:
-        mic.stop(blocking=False)
-        sink.join()
-        logging.info("recording stopped")
-
-    chunks = [f["chunk"] for f in sink.list]
-
-    assert len(chunks) > 0
 
 
 def _test_mic_to_wave_pipeline():
