+++ multisensor_pipeline/tests/image/test_pillow.py
