+++ multisensor_pipeline/tests/input/test_mouse.py
